+++ packages/ThemeEngine/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @memberjunction/theme-engine
 
+## 6.1.0-edge.3
+
 ## 6.1.0-edge.2
 
 ## 6.1.0-edge.1